apps/api/pyproject.toml
[project]
name = "win-prediction-api"
version = "0.1.0"
description = "FastAPI backend for Riot Win Prediction"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.19.0",
    "riotskillissue",
    "pandas>=2.2.0",
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
    "xgboost>=2.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
